AI: Perform the Extract Function of Teuchos::CommandLineProcessor::parse() using OpenHands
This refactoring was performed with OpenHands using OpenAI o3-mini using the
prompt:

----

Perform the Extract Function refactoring of the parent function
Teuchos::CommandLineProcessor::parse() in the file
packages/teuchos/core/src/Teuchos_CommandLineProcessor.cpp to make more
understandable and maintainable.

----

I had to tell it to also update the header file (it only updated the source file
at first).

It was not given any tools to compile so it had no way to test this.  It kept
asking me to check it.

There is zero chance this will compile.

This was not really agent mode at all.