Codex + GPT+5 extract funtion refactoring on parse()
This was performed in one prompt using the single prompt:

----

Perform the Extract Function refactoring on the function
Teuchos::CommandLineProcessor::parse() that is declared and defined in the
files packages/teuchos/core/src/Teuchos_CommandLineProcessor.hpp and
packages/teuchos/core/src/Teuchos_CommandLineProcessor.cpp.  Extract out
functions to make the code similer and easier to maintain. Update the parent
function to call the new funtions.

----

I did not try to build this but I could.  But this looks very good.